openai_gateway/pyproject.toml
[tool.poetry]
name = "openai-gateway"
version = "0.1.0"
description = "Secure OpenAI proxy gateway for Salesforce AI Email Assistant"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
httpx = "^0.25.2"
aioredis = "^2.0.1"
presidio-analyzer = "^2.2.33"
presidio-anonymizer = "^2.2.33"
pydantic = {extras = ["dotenv"], version = "^2.5.0"}
structlog = "^23.2.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.6"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
respx = "^0.20.1"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
bandit = "^1.7.5"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.bandit]
exclude_dirs = ["tests"] 